backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
google-genai==0.3.0
firebase-admin==6.4.0
pydantic==2.6.3
numpy==1.26.4
python-dotenv==1.0.1
